requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
langchain_google_genai
ddgs
duckduckgo-search
